API change: pass userVisibleOnly:true to PushManager.subscribe
Setting gcm_user_visible_only in the manifest is deprecated now that we
have standardized a cleaner alternative (in Chrome 44 we'll show a
warning, in Chrome 45 we'll drop support). See intent to ship thread[1].

Instead, when registering, pass an additional dictionary with a
userVisibleOnly key, specifically:
    serviceWorkerRegistration.pushManager.subscribe({
        userVisibleOnly: true
    }).then(...);
Still need gcm_user_visible_only in the manifest whilst Chrome 43 or 42
remain supported by the website.

[1]: https://groups.google.com/a/chromium.org/d/topic/blink-dev/_2JV9BTB1Ec/discussion