API change: GCM subscription ID is now included in endpoint
The PushSubscription objects returned by the Push API are going to stop
including subscriptionId (to better match the future IETF webpush
server-side protocol). In Chrome 44+ the endpoint has the subscriptionId
appended to it (separated by a "/").

In Chrome 44 only, to ease the transition, PushSubscription will have
the subscriptionId both appended to the endpoint and as a separate
property. See intent to deprecate thread[1] for details. For backwards
compatibility, the following logic should be used on the client side:

serviceWorkerRegistration.pushManager.subscribe(...).then(function(ps) {
    var endpoint = ps.endpoint;
    if ('subscriptionId' in ps &&
        !endpoint.includes(ps.subscriptionId))
      endpoint += "/" + ps.subscriptionId;

    // Now send |endpoint| to server rather than |ps.subscriptionId|,
    // and update server to parse the subscription ID from the endpoint.
});

Then on the server-side you'll need to strip the leading
"https://android.googleapis.com/gcm/send/" (note trailing slash) from
the endpoint in order to get the registration ID that you then pass to
GCM as usual.

[1]: https://groups.google.com/a/chromium.org/d/topic/blink-dev/CK13omVO5ds/discussion